refactor(frontend): collapse dual repositories barrel to single canonical (#11678)
